+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Minor
 * Avatar / GroupAvatar: make outline configurable(#173)
+* Masonry: Update non-virtualized Masonry to render all items regardless of the window
 * ExperimentalMasonry: remove component (#183)
 * Internal: Add flow-typed files for third party packages (#174)
 * Internal: Remove unused linter suppressions (#180)